+++ flask_express/response.py
@@ -350,7 +350,7 @@ class Response(ResponseBase):
                 filename = req.query.filename
                 return res.attachment(file_name)
             
-        ----- version-0.1.3 changes ------
+        ----- version-0.1.4 changes ------
         Now you can set the downloadable name of the attachment.
         :for example::